[Dialogs] Rollback of PR #9877 to fix the message text disappearing on device rotation when voiceover is active.
Note that only the change to MDCAlertController.m is rolled back, since the test method in MDCAlertControllerActionsTests.m has already changed since the PR.

The goal of this PR is to fix the Detail/message text that disappears when voiceover is active. It also fixes Meet's Dialog turning blank on device rotation.

# Explanation
`preferredContentSize` is calculated incorrectly in `viewWillLayoutSubviews`, causing the logic reverted in cl/302659116, which is meant only for extra large dialogs, to be executed for all dialogs. When executed for non-extra large dialogs, it results in the text disappearing on rotation when voiceover is active.

Removing the logic incidentally fixed the issue, but potentially left many other unresolved issues. Correcting the size calculations fixes the root cause, and while fixing this issue, also addresses the other potential issues that may result from incorrect size calculations.

PiperOrigin-RevId: 303316111