[Dialogs] Support custom alignment for the title image (independent of title)
Adding support for customizing the alignment of the title icon, independently of the alignment of the title.

To preserve backward compatibility, the default title-icon alignment is the same as the title. We're using the internal _alignIconWithTitle flag to track if the alignment had been implicitly assigned. Once the value is implicitly assigned, the title icon alignment is independent from the title.

## Issues
Closes #9545, #9824.

PiperOrigin-RevId: 303748434