[AppBar] Crash fix for VoiceOver + ContentInset
When VoiceOver is on, there currently is a contentOffset "pinning" added to the FlexibleHeader, this is because when voiceover is turned on, it would move the contentOffset undesirably and try to center the flexible header.

However, pinning the contentOffset to maximumHeightWithTopSafeArea would cause a crash in cases where that contentOffset value was invalid.

It can be an invalid value in cases where a contentInset is set to some different value on the scrollView, and/or the contentSize height of the content was very small. Strictly speaking, when the scrollView needs to re-set it's contentOffset again internally, that is a strong indicator that the contentOffset provided was invalid.

The crash occurred because the contentOffset pinning is triggered when the content offset changes. However, when the contentOffset is invalid, the scrollView will then call setContentOffset again to try and fix it to a correct value. This will in turn re-call the contentOffset pinning, causing an endless "livelock-like recursion". The reason it is not a traditional infinite recursion is because the contentOffset keeps changing at the next run loop, and it doesn't occur all in the same call stack.

To fix this, we made sure that the contentOffset number is valid by taking the topInset into regards and not only the provided maximum height.

PiperOrigin-RevId: 304136582