[Dialogs] Size justified-aligned title-icons to fit the alert's width.
To allow a meaningful contentMode setting for the titleIcon image view, the title icon image is sized to fit the alert's width. The height is then adjusted proportionally.

To maintain backward compatibility (and minimize client UI changes), the icon frame size is not adjusted for non-justified alignments.

This is a potential breaking change for clients that use Justified title, however it is unlikely that many clients (if any) would do that, since it's in most cases a bad UI/UX.

PiperOrigin-RevId: 304213073