[AppBar] Fix AppBar from jumping when contentOffset updates when VoiceOver is on.
The problem occurs because the transform of the flexible header view's Y positioning based on the contentOffset is not done in coordination with the scroll view animation.

Similarly to the jumping effect in https://github.com/material-components/material-components-ios/pull/9872 , we coordinate the transform when VoiceOver is on with the scroll view content offset change.

This was tested with the Expandable Cells example and the glitch is fixed.

PiperOrigin-RevId: 304352119