[Banner] Make the primary demo presentable.
The primary demo was previously not presentable, meaning it was not appearing in the Banner list of examples and one of the other examples was being selected as the "primary" example. This was breaking testing automation because the expectation for each component is that it has a primary demo and that it is presentable.

PiperOrigin-RevId: 306297190