[AppBar] Round content offset clamp to nearest integer.
An issue was found regarding UIKit re-setting the content offset when the value is a fractional number due to screen densities. By rounding the offset clamp to the nearest integer we are assuring the contentOffset value will be valid and UIKit would not try to alter it causing an infinite loop and thus a crash.

PiperOrigin-RevId: 307810013