[MDC/ActionSheet] Support setting image later in lifecycle
This allows clients to set the image property on an MDCActionSheetAction later in the lifecycle of the action. Particularly after an ActionSheet has been presented. We only set the `actionSheet` property in `viewWillAppear` since clients can use the same action on multiple action sheets. This change will make it so that the action sheet on screen will be updated.

Alternatively I could have done KVO but I felt that was more overhead than just having a custom setter.

PiperOrigin-RevId: 309470766