[FlexibleHeader] Extract a shifter logical unit.
This logical unit will house all logic related to the shift behavior. It is being extracted as a distinct logical unit so that it can be better unit tested and validated in terms of behavior. This change introduces an initial test for the basic skeletal object accordingly.

In subsequent changes, we will move logic from MDCFlexibleHeaderView into this new object.

PiperOrigin-RevId: 311203197