[NavigationDrawer] Fix bug where setting `maximumDrawerHeight` and `maximumInitialDrawerHeight` would allow scrolling.
If a client sets the `maximumDrawerHeight` and the `maximumInitialDrawerHeight` to the same value then the drawer should be opened to its _max_ height. Currently if that happens the drawer allows for _safeArea.top_ pts of scrolling. This is due to us adjusting the `contentHeaderTopInset` which is then used to layout all the content. We typically adjust it because we assume that the header adds the _safeArea.top_ pts to it's height as it approaches the safe area. So really the `contentHeaderTopInset` typically _maxes out_ at that amount but if we never approach the safe area then we need to allow for this value to be 0.

Tested in the Catalog via Infinite Scrolling example.

PiperOrigin-RevId: 313814020