[NavigationDrawer] Fix bug where `maximumDrawerHeight` wouldn't allow for the correct height.
In https://github.com/material-components/material-components-ios/commit/33daf325e650b27a44004ff44b97a61d98221b79#diff-a3db687b7d16849448cb7f189976eb0f we added this property to allow the drawer to support max height. In adding this property my assumption was that we needed to shift the scroll view down by the amount of the topSafeArea but this is an incorrect assumption, we should leave it to the client to determine the height of the drawer an not _magically_ shift it down by an amount. Also, we tied the drawer max height to the `shouldIncludeSafeAreaInContentHeight` to the height of the drawer this caused the drawer to be shifted down by the safeArea.bottom height. The property `shouldIncludeSafeAreaInContentHeight` is related to the `maximumInitialDrawerHeight` property but not the `maximumDrawerHeight` property. Both of these values should be on the client to pass in instead of us _magically_ shifting down potentially causing issues where a client expects the drawer to be "x" height but it shows as "x - (safeArea.top + safeArea.bottom)".

Tested via the catalog examples.

PiperOrigin-RevId: 314127239