[TextControls] Add notion of horizontal positioning reference
No visible or public API changes. Internal refactor.

This change adds the notion of a horizontal positioning reference. This will make it so that different styles can have different horizontal padding values.

Prior to this change, there was only one "horizontal padding" value. This change separates it into two values: one for the padding between leftmost and rightmost views and the left and right edges of the TextControl, and one for the padding between subviews of the TextControl. These two values live on a horizontal positioning reference object. Different styles will one day have different horizontal positioning reference objects.

For now only TextControl text fields will use this.
It's a lot of files but trivial changes in most, if not all, of them.

PiperOrigin-RevId: 314748534