[TextControls] Add padding customizing properties to MDCBaseTextField and MDCBaseTextArea
This CL adds two new properties, leadingEdgePaddingOverride and trailingEdgePaddingOverride, to MDCBaseTextField and MDCBaseTextArea. When these properties are set, they will be used as padding values by the text field. When they are not set, the text field will use default values. I decided to make these properties NSNumbers, so there is a distinction between setting them to a @0 and leaving them as nil. This is where optionals would come in handy...

Things to come in upcoming CLs:
* Snapshot tests for text areas
* Will rename MDCTextControlVerticalPositioningReference to MDCTextControlVerticalPositioning, for consistency with its horizontal equivalent

PiperOrigin-RevId: 317349272