Update sizeThatFits calculation to include layout margins in the height calculation and ensure entire leading button can be clicked when trailing button is hidden, by constraining the buttonContainerView to always be at least as tall as the leading button.
Note: With the update in the height calculation, the additional constraint seems to have no effect, and the banner's layout is correct either way. I prefer to keep it, however, as it'll ensure we don't get the same "button cannot be clicked" error in the future, at least not without breaking constraints. Happy to reconsider this, though!
PiperOrigin-RevId: 321214409