[Shapes] Updates the shapeGenerator with a line width to inset the line rather than center it.
There is a visual difference when using lineWidth instead of borderWidth due to lineWidth drawing on the path, and borderWidth insetting from the path.
To resolve this, the path is made smaller by half the width of the line, thus to create a similar visual appearance as if borderWidth was set.

PiperOrigin-RevId: 322615641