[Shapes] Fix divide by zero bug.
This bug was happening when an MDCShapedShadowLayer has an empty frame and a positive shapedBorderWidth value, and a path is assigned. The MDCShapedShadowLayer was attempting to calculate a value in relation to the bounds of the MDCShapedShadowLayer, which in turn was resulting in a divide by zero.

This change adds safeguards around the divisions to ensure that a division does not occur if the width or height are sufficiently close to zero.

PiperOrigin-RevId: 322768484