[TextControls] Add KVO for assistive labels
This change makes it so that clients will not have to explicitly trigger a layout pass and resize text fields or text areas when they set assistive label text in Auto Layout.
For Manual Layout, they will not have to trigger a layout pass, but they will still have to resize the view.

PiperOrigin-RevId: 324873963