Automated g4 rollback of changelist 331153646.
*** Reason for rollback ***

This is causing regression b/168438879, when receiving .screenchanged notification, we switch focus to other element on the page, and this cause the snackbar to auto dismiss.

*** Original change description ***

Enable shouldDismissOnOverlayTap for VO enabled scenario

BEGIN_PUBLIC

Enable shouldDismissOnOverlayTap for VO enabled scenario

END_PUBLIC

Per discussion in b/167609843, based on GAR requirement, snackbar with action button shouldn't self-dismiss when VO is enabled, but If a user does an action then we should/could dismiss the snackbar. This CL uses shouldDismissOnOverlayTap property added in cl/310921840 to enable the dismiss when user takes action on overlay.

***

PiperOrigin-RevId: 331654884