[Dialogs] Add an accessoryView with a collection example
This example demonstrates adding a collection view as an accessory view of a Material alert.  It also demonstrates adding a horizontal line above the actions, as shown on material.io's spec for Confirmation Dialogs. The line will be fixed to reach the edges of the dialogs in a followup PR (cl/331128888).

PiperOrigin-RevId: 333376220