[Manage People/Links] Removed a condition for `presentingViewYOffset` in `MDCBottomDrawerContainerViewController`
BEGIN_PUBLIC
Updated the `presentingViewYOffset` conditional in `MDCBottomDrawerContainerViewController`. Checking the `verticalSizeClass` trait doesn't seem to provide value for the conditional seeing as regardless of screen size we would still need a y offset and would default to zero if the value wasn't large enough anyways. Taking off this conditional would also make it easier to create unit tests for a future related CL that will take the `presentingViewYOffset` into account. I tested this change by running the drive app and the material catalog app to search for an UX discrepancies. I also ran the unit and snapshot tests associated with this class and they passed fine with just a small change in the golden image needed
END_PUBLIC

PiperOrigin-RevId: 345987236