[Manage People/Links] Added `presentingViewYOffset` to `contentHeightSurplus` in `MDCBottomDrawerContainerViewController`
BEGIN_PUBLIC
Updated the `contentHeightSurplus` method to add in an additional value of `presentingViewYOffset'. In the current iteration the `presentingViewYOffset' gets cancelled out by being included in both `scrollingDistance` and `containerHeight`. This change will take that offset back in to account and resolves a UX bug that we're currently seeing devices with smaller screen sizes using the PageSheet presentation style (b/173830907).
END_PUBLIC

PiperOrigin-RevId: 347036858