[TextControls] Prevent duplicate accessibility label and value audio feedback when placeholder is equal to the label.
Voice Over reads both the accessibility label and the accessibility value in succession.
If no value is set on the text field, the accessibility value will be the placeholder. This means that if a placeholder is set and the value is the same as the label Voice Over will read the same string twice. Ex, "Label, Label, Text Field". This is confusing to the user, so if we determine this to be case, we manually return an empty accessibility value.

PiperOrigin-RevId: 387372773