[BottomNavigation] Remove unnecessary NSNull check.
v45.0.0 (cl/185714778) of MDC introduced the patch-fix in the item view, and v85.9.1 introduced a fix to the root cause (cl/259949781) which was the KVO router not properly casting NSNull to nil. The root cause fix in v85.9.1 removes the need for the check in the item view, which is otherwise confusing given the signature of the method.

PiperOrigin-RevId: 415599493