[Badges] Add support for borders.
Borders are expected to be drawn outside the badge, but CALayer's border behavior drawers the borders inside of the view. Additionally, the border color is a CGColorRef meaning it won't react automatically to trait collection changes.

This change introduces two new APIs to support outer borders:

1. borderWidth
2. borderColor

`borderWidth` is a proxy for `self.layer.borderWidth`; the only difference between setting `self.borderWidth` and `self.layer.borderWidth` is that `self.borderWidth`'s setter will invalidate the layout and intrinsic content size. This is documented accordingly in API docs.

`borderColor` is a `UIColor`, which is stored adjacent to `self.layer.borderColor`. `self.layer.borderColor` can still technically be set directly, but the value will be reset to whatever `self.borderColor` is any time the trait collection changes.

PiperOrigin-RevId: 416155619