[BottomNavigation] Replace null_resettable with nonnull.
null_resettable is imported to Swift as an optional type, which requires awkward nil checks when working with the appearance properties, even though the return value is guaranteed to be non-nil.

Instead, appearance properties are now all specified as nonnull. This allows Swift source to use the appearance APIs in a much simpler fashion.

To achieve the same "null_resettable" behavior as before, simply assign an empty instance of an appearance object.

PiperOrigin-RevId: 418678212