[BottomNavigation] Tighten up selection indicator animations.
This change moves the selection animation logic into the item view so that it we can remove view manipulation logic from the parent bottom navigation view. This change is possible because we do not need to slide selection indicators between item views.

In this change, the animation for the selection indicator has been tuned with the following motion design considerations:

1. The deselected item never animates. This avoids excess motion noise when transitioning to the new selection.
2. The selected item animates its selection indicator separately from its image. This allows us to give greater emphasis to the selection indicator's animation by making the image animation happen faster.

PiperOrigin-RevId: 418889483