Fix text color when reseting items while maintaining the selected tab.
If tab bar items are updated while there is already a selected item the early exit path in `setSelectedItem:` will be taken. This will bypass styling for the selected item (which will have a corresponding newly constructed MDCTabBarViewItemView that needs styling). This results in the MDCTabBarViewItemView's textColor being incorrect.

The fix proposed in this CL resets the selectedItem to nil so that the early exit check fails and allows the tab to be themed.

PiperOrigin-RevId: 440864525