[BottomNavigation] Fix bug where the badge color was not reacting to the background color set in the appearance property.
This bug was caused by an attempt to maintain backward compatibility with the legacy behavior of setting badgeColor, which treated nil as clearColor. This logic was causing the default badge color (which is red) to always override the appearance background color.

This change simply removes that override logic, because there shouldn't be a case where a badge's background color is set to transparent because doing so would negatively affect legibility of the badge.

PiperOrigin-RevId: 466780235