[M3CButton] Added imageEdgeInsetsWithImageAndTitle property.
This will allow the internal logic to set the imageEdgeInsets when there is image and text as well as zero it out safely when no text or image are present.

Prior to this fix: if no image or text were set when the branding API is called the imageEdgeInset may get overwritten with zeros.

PiperOrigin-RevId: 494337628