[Chips] Allow chips to be selected in MDCChipField.
Prior to this change, MDCChipField's chips would respond to tap gestures by showing a highlight state but the chips would not become selected. This meant that it was not possible to tap a chip and then delete it by pressing the backspace key, as is the expected behavior.

This change introduces a tap gesture onto each chip view that is added to the field. When a chip view is tapped, the chip view's selection state is toggled.

After this change, the user will now be able to select one or more chips and delete them by pressing the backspace button.

PiperOrigin-RevId: 626041379